Moved to an different License scheme
Licenses are now defined as an yaml array in the header.
The key ist license and the first element in the array is
the name of the license while the second is the URL.
This change does not support licenses with only the license
name (no URL). This may come later.